@@ -11,27 +11,24 @@
     "node": ">=8.9"
   },
   "scripts": {
-    "build:apidocs": "lb-apidocs",
-    "build": "dotenv-load lb-tsc --outDir dist",
-    "build:watch": "dotenv-load lb-tsc --watch",
     "build:docker": "dotenv-load node ./bin/docker-build.js",
+    "build:watch": "dotenv-load lb-tsc --watch",
+    "build": "dotenv-load lb-tsc --outDir dist",
     "clean": "lb-clean dist",
-    "lint": "npm run prettier:check && npm run tslint",
-    "lint:fix": "npm run tslint:fix && npm run prettier:fix",
-    "prettier:cli": "lb-prettier \"**/*.ts\" \"**/*.js\"",
-    "prettier:check": "npm run prettier:cli",
-    "prettier:fix": "npm run prettier:cli -- --write",
-    "tslint": "lb-tslint",
-    "tslint:fix": "npm run tslint -- --fix",
-    "pretest": "npm run clean && npm run build",
-    "test": "dotenv-load lb-mocha --allow-console-logs \"dist/test\"",
-    "posttest": "npm run lint",
-    "test:dev": "lb-mocha --allow-console-logs dist/test/**/*.js && npm run posttest",
-    "prestart": "npm run build",
-    "start": "dotenv-load node .",
-    "prepublishOnly": "npm run test",
-    "deploy": "dotenv-load npm run build && npm run deploy:docker",
     "deploy:docker": "npm run build:docker && dotenv-load node ./bin/docker-deploy.js",
+    "deploy": "dotenv-load npm run build && npm run deploy:docker",
+    "eslint:fix": "npm run eslint -- --fix",
+    "eslint": "eslint --report-unused-disable-directives bin/**/*.[tj]s src/**/*.[tj]s test/**/*.[tj]s",
+    "lint:fix": "npm run prettier:fix && npm run eslint:fix",
+    "lint": "npm run prettier:check && npm run eslint",
+    "posttest": "npm run lint",
+    "prepublishOnly": "npm run test",
+    "pretest": "npm run clean && npm run build",
+    "prettier:check": "npm run prettier:cli -- -l",
+    "prettier:cli": "lb-prettier \"**/*.ts\" \"**/*.js\"",
+    "prettier:fix": "npm run prettier:cli -- --write",
+    "start": "dotenv-load npm run build && dotenv-load node .",
+    "test": "dotenv-load lb-mocha \"dist/test\" --allow-console-logs",
     "version:major": "npm version --no-git-tag-version major",
     "version:minor": "npm version --no-git-tag-version minor",
     "version:patch": "npm version --no-git-tag-version patch"
@@ -75,7 +72,8 @@
     "uuid": "^3.4.0"
   },
   "devDependencies": {
-    "@loopback/build": "^2.1.0",
+    "@loopback/build": "^4.0.0",
+    "@loopback/eslint-config": "^6.0.0",
     "@loopback/testlab": "^1.10.3",
     "@types/assert": "^1.4.6",
     "@types/node": "^10.17.16",
@@ -85,6 +83,16 @@
     "dotenv-load": "^2.0.0",
     "shelljs": "^0.8.3",
     "tslint": "^5.20.1",
-    "typescript": "^3.7.5"
+    "typescript": "^3.8.3"
+  },
+  "eslintConfig": {
+    "extends": "@loopback/eslint-config",
+    "rules": {
+      "@typescript-eslint/camelcase": "off",
+      "@typescript-eslint/no-explicit-any": "off",
+      "@typescript-eslint/class-name-casing": "off",
+      "@typescript-eslint/interface-name-prefix": "off",
+      "no-empty": "off"
+    }
   }
 }
