@@ -1,17 +1,15 @@
-import { Signature } from '../entities';
-import { TypeORMDataSource } from '../datasources';
-import { inject } from '@loopback/core';
-import { TypeORMRepository } from './typeorm-repository';
+import {Signature} from '../entities';
+import {TypeORMDataSource} from '../datasources';
+import {inject} from '@loopback/core';
+import {TypeORMRepository} from './typeorm-repository';
 
 export class SignatureRepository extends TypeORMRepository<
   Signature,
   typeof Signature.prototype.id
-  >  {
-  dataSource: TypeORMDataSource
+> {
+  dataSource: TypeORMDataSource;
 
-  constructor(
-    @inject('datasources.typeorm') dataSource: TypeORMDataSource,
-  ) {
+  constructor(@inject('datasources.typeorm') dataSource: TypeORMDataSource) {
     super(Signature, dataSource);
   }
 }
