@@ -1,14 +1,23 @@
-import { api, post, param, RestBindings, RequestContext, get, Response } from "@loopback/rest";
-import { authenticate, AuthenticationBindings } from "@loopback/authentication";
-import { inject } from "@loopback/core";
-import { UserProfile } from "../models";
-import { TypeORMDataSource } from "../datasources";
-import { TypeORMRepository } from "../repositories";
-import { IGenericRepository, IGenericEntity } from "../generic-controllers/generic.controller";
+import {
+  api,
+  param,
+  RestBindings,
+  RequestContext,
+  get,
+  Response,
+} from '@loopback/rest';
+import {authenticate, AuthenticationBindings} from '@loopback/authentication';
+import {inject} from '@loopback/core';
+import {UserProfile} from '../models';
+import {TypeORMDataSource} from '../datasources';
+import {
+  IGenericRepository,
+  IGenericEntity,
+} from '../generic-controllers/generic.controller';
 
 @api({
   basePath: `${process.env.PREFIX}/optimize`,
-  paths: {}
+  paths: {},
 })
 class OptimizationController {
   constructor(
@@ -16,7 +25,7 @@ class OptimizationController {
     @inject('datasources.typeorm') private dataSource: TypeORMDataSource,
     @inject(RestBindings.Http.RESPONSE) private response: Response,
     @inject.context() private ctx: RequestContext,
-  ) { }
+  ) {}
 
   @authenticate('OPTIMIZE')
   @get('/refresh', {
@@ -25,13 +34,11 @@ class OptimizationController {
     responses: {
       '200': {
         description: 'Refresh materialized views',
-      }
-    }
+      },
+    },
   })
-  async refresh(
-    @param.query.string('view') view?: string
-  ): Promise<void> {
-    await this.dataSource.refresh_materialized_views(view)
+  async refresh(@param.query.string('view') view?: string): Promise<void> {
+    await this.dataSource.refresh_materialized_views(view);
   }
 
   @authenticate('OPTIMIZE')
@@ -41,17 +48,17 @@ class OptimizationController {
     responses: {
       '200': {
         description: 'Ensure index is present on field `table_name.deep.field`',
-      }
-    }
+      },
+    },
   })
-  async index(
-    @param.query.string('field') field: string
-  ): Promise<void> {
-    const field_split = field.split('.')
-    const table = field_split[0]
-    const repo = await this.ctx.get<IGenericRepository<IGenericEntity>>(`repositories.${table}`)
-    await repo.ensureIndex(field_split.slice(1).join('.'))
+  async index(@param.query.string('field') field: string): Promise<void> {
+    const field_split = field.split('.');
+    const table = field_split[0];
+    const repo = await this.ctx.get<IGenericRepository<IGenericEntity>>(
+      `repositories.${table}`,
+    );
+    await repo.ensureIndex(field_split.slice(1).join('.'));
   }
 }
 
-export { OptimizationController }
+export {OptimizationController};
