@@ -1,13 +1,19 @@
-import { Entity as TypeORMEntity, Column, Generated, OneToMany, Index } from 'typeorm';
-import { Entity as LBEntity, model, property } from '@loopback/repository';
-import { getJsonSchema } from '@loopback/rest';
-import { Library } from './library.model';
+import {
+  Entity as TypeORMEntity,
+  Column,
+  Generated,
+  OneToMany,
+  Index,
+} from 'typeorm';
+import {Entity as LBEntity, model, property} from '@loopback/repository';
+import {getJsonSchema} from '@loopback/rest';
+import {Library} from './library.model';
 
 @model({
   name: 'Resource',
   description: 'A table for storing validatable resources in the database',
   settings: {
-    strict: false
+    strict: false,
   },
 })
 @TypeORMEntity({
@@ -21,7 +27,7 @@ export class Resource extends LBEntity {
     unique: true,
   })
   @Generated()
-  _id: number
+  _id: number;
 
   @property({
     type: 'string',
@@ -41,20 +47,23 @@ export class Resource extends LBEntity {
     required: true,
     default: {},
   })
-  @Index('resources_meta_gin_index', { synchronize: false })
-  @Index('resources_meta_gist_fts_index', { synchronize: false })
+  @Index('resources_meta_gin_index', {synchronize: false})
+  @Index('resources_meta_gist_fts_index', {synchronize: false})
   @Column({
     type: 'jsonb',
   })
   meta: {
-    [key: string]: any
+    [key: string]: any;
   };
 
-  @OneToMany(type => Library, library => library.resource)
+  @OneToMany(
+    type => Library,
+    library => library.resource,
+  )
   _libraries: Promise<Library[]>;
 
   constructor(data?: Partial<Resource>) {
     super(data);
   }
 }
-export const ResourceSchema = getJsonSchema(Resource)
+export const ResourceSchema = getJsonSchema(Resource);
