@@ -1,12 +1,22 @@
-import { Resource as ResourceEntity, ResourceSchema, Library, Signature } from '../entities';
-import { ResourceRepository } from '../repositories';
-import { GenericControllerFactory } from './generic.controller'
-import { getFilterSchemaFor, get, param, getWhereSchemaFor } from '@loopback/rest';
-import { Library as LibraryController } from './library.controller';
-import { Signature as SignatureController } from './signature.controller';
-import { Filter, Count, Where, Fields } from '@loopback/repository';
-import { authenticate } from '@loopback/authentication';
-import { inject } from '@loopback/core';
+import {
+  Resource as ResourceEntity,
+  ResourceSchema,
+  Library,
+  Signature,
+} from '../entities';
+import {ResourceRepository} from '../repositories';
+import {GenericControllerFactory} from './generic.controller';
+import {
+  getFilterSchemaFor,
+  get,
+  param,
+  getWhereSchemaFor,
+} from '@loopback/rest';
+import {Library as LibraryController} from './library.controller';
+import {Signature as SignatureController} from './signature.controller';
+import {Filter, Count, Where, Fields} from '@loopback/repository';
+import {authenticate} from '@loopback/authentication';
+import {inject} from '@loopback/core';
 
 const GenericResourceController = GenericControllerFactory<
   ResourceEntity,
@@ -17,7 +27,7 @@ const GenericResourceController = GenericControllerFactory<
   GenericEntitySchema: ResourceSchema,
   modelName: 'Resource',
   basePath: `${process.env.PREFIX}/resources`,
-})
+});
 
 export class Resource extends GenericResourceController {
   @authenticate('GET.resources.libraries')
@@ -25,25 +35,25 @@ export class Resource extends GenericResourceController {
   async libraries(
     @inject('controllers.Library') libraryController: LibraryController,
     @param.path.string('id') id: string,
-    @param.query.object('filter', getFilterSchemaFor(Library)) filter?: Filter<Library>,
-    @param.query.string('filter_str') filter_str: string = '',
-    @param.query.boolean('contentRange') contentRange: boolean = true,
+    @param.query.object('filter', getFilterSchemaFor(Library))
+    filter?: Filter<Library>,
+    @param.query.string('filter_str') filter_str = '',
+    @param.query.boolean('contentRange') contentRange = true,
   ): Promise<Library[]> {
-    if (filter_str !== '' && filter == null)
-      filter = JSON.parse(filter_str)
+    if (filter_str !== '' && filter == null) filter = JSON.parse(filter_str);
 
     const libraries = await libraryController.find({
       filter: {
-        ...(filter || {}),
+        ...(filter ?? {}),
         where: {
-          ...((filter || {}).where || {}),
-          resource: id
-        }
+          ...((filter ?? {}).where ?? {}),
+          resource: id,
+        },
       },
       contentRange,
-    })
+    });
 
-    return libraries
+    return libraries;
   }
 
   @authenticate('GET.resources.libraries.count')
@@ -51,20 +61,18 @@ export class Resource extends GenericResourceController {
   async libraries_count(
     @inject('controllers.Library') libraryController: LibraryController,
     @param.path.string('id') id: string,
-    @param.query.object('where', getWhereSchemaFor(Library)) where?: Where<Library>,
-    @param.query.string('where_str') where_str: string = '',
+    @param.query.object('where', getWhereSchemaFor(Library))
+    where?: Where<Library>,
+    @param.query.string('where_str') where_str = '',
   ): Promise<Count> {
-    if (where_str !== '' && where == null)
-      where = JSON.parse(where_str)
+    if (where_str !== '' && where == null) where = JSON.parse(where_str);
 
-    const count = await libraryController.count(
-      {
-        ...(where || {}),
-        resource: id
-      },
-    )
+    const count = await libraryController.count({
+      ...(where ?? {}),
+      resource: id,
+    });
 
-    return count
+    return count;
   }
 
   @authenticate('GET.resources.signatures')
@@ -73,28 +81,28 @@ export class Resource extends GenericResourceController {
     @inject('controllers.Library') libraryController: LibraryController,
     @inject('controllers.Signature') signatureController: SignatureController,
     @param.path.string('id') id: string,
-    @param.query.object('filter', getFilterSchemaFor(Signature)) filter?: Filter<Signature>,
-    @param.query.string('filter_str') filter_str: string = '',
-    @param.query.boolean('contentRange') contentRange: boolean = true,
+    @param.query.object('filter', getFilterSchemaFor(Signature))
+    filter?: Filter<Signature>,
+    @param.query.string('filter_str') filter_str = '',
+    @param.query.boolean('contentRange') contentRange = true,
   ): Promise<Signature[]> {
-    if (filter_str !== '' && filter == null)
-      filter = JSON.parse(filter_str)
+    if (filter_str !== '' && filter == null) filter = JSON.parse(filter_str);
     // TODO: use ORM inner join
     const libraries = await libraryController.find({
       filter: {
-        where: { resource: id },
+        where: {resource: id},
         fields: ['id'] as Fields<Library>,
       },
-    })
-    return await signatureController.find({
+    });
+    return signatureController.find({
       filter: {
         ...filter,
         where: {
-          ...((filter || {}).where || {}),
-          library: { inq: libraries.map(({ id }) => id ) },
+          ...((filter ?? {}).where ?? {}),
+          library: {inq: libraries.map(({id: _id}) => _id)},
         },
-      }
-    })
+      },
+    });
   }
 
   @authenticate('GET.resources.signatures.count')
@@ -103,30 +111,30 @@ export class Resource extends GenericResourceController {
     @inject('controllers.Library') libraryController: LibraryController,
     @inject('controllers.Signature') signatureController: SignatureController,
     @param.path.string('id') id: string,
-    @param.query.object('where', getWhereSchemaFor(Library)) where?: Where<Library>,
-    @param.query.string('where_str') where_str: string = '',
+    @param.query.object('where', getWhereSchemaFor(Library))
+    where?: Where<Library>,
+    @param.query.string('where_str') where_str = '',
   ): Promise<Count> {
-    if (where_str !== '' && where == null)
-      where = JSON.parse(where_str)
-      
-    let count = 0
+    if (where_str !== '' && where == null) where = JSON.parse(where_str);
+
+    let count = 0;
     for (const library of await libraryController.find({
       filter: {
         where: {
-          resource: id
+          resource: id,
         },
-        fields: [ 'id' ]
-      }
+        fields: ['id'],
+      },
     } as any)) {
       count += (
         await libraryController.signatures_count(
           signatureController,
           library.id,
-          where || {},
+          where ?? {},
         )
-      ).count
+      ).count;
     }
 
-    return { count }
+    return {count};
   }
 }
