@@ -1,9 +1,9 @@
-import { App } from './application';
-import { ApplicationConfig } from '@loopback/core';
-import { UserProfileRepository } from './repositories';
-import { TypeORMDataSource } from './datasources';
+import {App} from './application';
+import {ApplicationConfig} from '@loopback/core';
+import {UserProfileRepository} from './repositories';
+import {TypeORMDataSource} from './datasources';
 
-export { App };
+export {App};
 
 export async function main(options: ApplicationConfig = {}) {
   const app = new App({
@@ -11,8 +11,14 @@ export async function main(options: ApplicationConfig = {}) {
       openApiSpec: {
         setServersFromRequest: true,
         endpointMapping: {
-          [`${process.env.PREFIX}/openapi.json`]: { version: '3.0.0', format: 'json' },
-          [`${process.env.PREFIX}/openapi.yml`]: { version: '3.0.0', format: 'yaml' },
+          [`${process.env.PREFIX}/openapi.json`]: {
+            version: '3.0.0',
+            format: 'json',
+          },
+          [`${process.env.PREFIX}/openapi.yml`]: {
+            version: '3.0.0',
+            format: 'yaml',
+          },
         },
       },
       expressSettings: {
@@ -24,10 +30,10 @@ export async function main(options: ApplicationConfig = {}) {
   await app.start();
 
   if (
-    process.env['ADMIN_USERNAME'] !== undefined
-    && process.env['ADMIN_PASSWORD'] !== undefined
+    process.env['ADMIN_USERNAME'] !== undefined &&
+    process.env['ADMIN_PASSWORD'] !== undefined
   ) {
-    const userRepo = await app.getRepository(UserProfileRepository)
+    const userRepo = await app.getRepository(UserProfileRepository);
     console.log('Creating admin...');
     await userRepo.deleteAll();
     await userRepo.create({
@@ -35,11 +41,11 @@ export async function main(options: ApplicationConfig = {}) {
       username: process.env['ADMIN_USERNAME'],
       password: process.env['ADMIN_PASSWORD'],
       roles: '^.+$',
-    })
+    });
   }
 
-  const typeorm = await app.get<TypeORMDataSource>('datasources.typeorm')
-  await typeorm.connect()
+  const typeorm = await app.get<TypeORMDataSource>('datasources.typeorm');
+  await typeorm.connect();
 
   const url = app.restServer.url;
   console.log(`Server is running at ${url}`);
