@@ -1,11 +1,11 @@
-import {MigrationInterface, QueryRunner} from "typeorm";
+import {MigrationInterface, QueryRunner} from 'typeorm';
 
-export class materializedViewWithLists1569597110370 implements MigrationInterface {
-
+export class materializedViewWithLists1569597110370
+  implements MigrationInterface {
   public async up(queryRunner: QueryRunner): Promise<any> {
     await queryRunner.query(`
       drop materialized view if exists libraries_key_value_counts;
-    `)
+    `);
     await queryRunner.query(`
       create materialized view libraries_key_value_counts
       as
@@ -19,11 +19,11 @@ export class materializedViewWithLists1569597110370 implements MigrationInterfac
         jsonb_deep_key_value(row_to_json(libraries)::jsonb) on true
       group by key, value
       order by count desc;
-    `)
+    `);
 
     await queryRunner.query(`
       drop materialized view if exists signatures_key_value_counts;
-    `)
+    `);
     await queryRunner.query(`
       create materialized view signatures_key_value_counts
       as
@@ -37,11 +37,11 @@ export class materializedViewWithLists1569597110370 implements MigrationInterfac
         jsonb_deep_key_value(row_to_json(signatures)::jsonb) on true
       group by key, value
       order by count desc;
-    `)
+    `);
 
     await queryRunner.query(`
       drop materialized view if exists entities_key_value_counts;
-    `)
+    `);
     await queryRunner.query(`
       create materialized view entities_key_value_counts
       as
@@ -55,11 +55,11 @@ export class materializedViewWithLists1569597110370 implements MigrationInterfac
         jsonb_deep_key_value(row_to_json(entities)::jsonb) on true
       group by key, value
       order by count desc;
-    `)
+    `);
 
     await queryRunner.query(`
       drop materialized view if exists libraries_signatures_key_value_counts;
-    `)
+    `);
     await queryRunner.query(`
       create materialized view libraries_signatures_key_value_counts
       as
@@ -76,22 +76,21 @@ export class materializedViewWithLists1569597110370 implements MigrationInterfac
           jsonb_deep_key_value(row_to_json(signatures)::jsonb) on true
       group by libraries.uuid, key, value
       order by count desc;
-    `)
+    `);
   }
 
   public async down(queryRunner: QueryRunner): Promise<any> {
     await queryRunner.query(`
       drop materialized view libraries_key_value_counts;
-    `)
+    `);
     await queryRunner.query(`
       drop materialized view signatures_key_value_counts;
-    `)
+    `);
     await queryRunner.query(`
       drop materialized view entities_key_value_counts;
-    `)
+    `);
     await queryRunner.query(`
       drop materialized view libraries_signatures_key_value_counts;
-    `)
+    `);
   }
-
 }
