@@ -1,6 +1,16 @@
-import { AuthenticateFn, AuthenticationBindings } from '@loopback/authentication';
-import { inject } from '@loopback/context';
-import { FindRoute, InvokeMethod, ParseParams, Reject, RequestContext, RestBindings, Send, SequenceHandler, ExternalExpressRoutes } from '@loopback/rest';
+import {AuthenticateFn, AuthenticationBindings} from '@loopback/authentication';
+import {inject} from '@loopback/context';
+import {
+  FindRoute,
+  InvokeMethod,
+  ParseParams,
+  Reject,
+  RequestContext,
+  RestBindings,
+  Send,
+  SequenceHandler,
+  ExternalExpressRoutes,
+} from '@loopback/rest';
 import debug from './util/debug';
 
 const SequenceActions = RestBindings.SequenceActions;
@@ -12,25 +22,25 @@ export class Sequence implements SequenceHandler {
     @inject(SequenceActions.INVOKE_METHOD) protected invoke: InvokeMethod,
     @inject(SequenceActions.SEND) public send: Send,
     @inject(SequenceActions.REJECT) public reject: Reject,
-    @inject(AuthenticationBindings.AUTH_ACTION) protected authenticateRequest: AuthenticateFn,
-  ) { }
+    @inject(AuthenticationBindings.AUTH_ACTION)
+    protected authenticateRequest: AuthenticateFn,
+  ) {}
 
   async handle(context: RequestContext) {
     try {
-      const start = Date.now()
-      const { request, response } = context;
+      const start = Date.now();
+      const {request, response} = context;
       const route = this.findRoute(request);
-      request.setTimeout(0, () => { });
+      request.setTimeout(0, () => {});
 
       // Hotfix from https://github.com/strongloop/loopback-next/issues/1144#issuecomment-438359985
       if (!(route instanceof ExternalExpressRoutes)) {
         if (request.headers === undefined) {
-          request.headers = {}
+          request.headers = {};
         }
         if (request.headers.authorization === undefined) {
-          request.headers.authorization = 'Basic ' + Buffer.from(
-            'guest:guest'
-          ).toString('base64')
+          request.headers.authorization =
+            'Basic ' + Buffer.from('guest:guest').toString('base64');
         }
 
         await this.authenticateRequest(request);
@@ -43,19 +53,19 @@ export class Sequence implements SequenceHandler {
       try {
         response.setHeader(
           'Access-Control-Expose-Headers',
-          [
-            ...response.getHeaderNames(),
-            'X-Duration',
-          ].join(',')
+          [...response.getHeaderNames(), 'X-Duration'].join(','),
         );
-        response.setHeader('X-Duration', JSON.stringify(Number(Date.now() - start) / 1000));
+        response.setHeader(
+          'X-Duration',
+          JSON.stringify(Number(Date.now() - start) / 1000),
+        );
       } catch (e) {
-        debug(e)
+        debug(e);
       }
 
       this.send(response, result);
     } catch (err) {
-      debug(err)
+      debug(err);
       this.reject(context, err);
     }
   }
