@@ -1,2 +1,2 @@
-import * as createDebug from 'debug'
-export default createDebug('signature-commons-metadata-api')
+import createDebug from 'debug';
+export default createDebug('signature-commons-metadata-api');
