@@ -2,10 +2,13 @@
 
 export function makeTemplate(
   templateString: string,
-  templateVariables: {[key: string]: string}
+  templateVariables: {[key: string]: string},
 ) {
   const keys = Object.keys(templateVariables);
   const values = Object.values(templateVariables);
-  let templateFunction = new Function(...keys, `return \`${templateString}\`;`);
+  const templateFunction = new Function(
+    ...keys,
+    `return \`${templateString}\`;`,
+  );
   return templateFunction(...values);
 }
