@@ -1,4 +1,4 @@
-import { ObjectItems } from './object-items'
+import {ObjectItems} from './object-items';
 import debug from '../util/debug';
 
 /**
@@ -7,19 +7,17 @@ import debug from '../util/debug';
  * @param objs The objects to enumerate the keys of
  * @param depth The depth to go into object keys
  */
-export function keyCounts(objs: any[], depth: number = 0) {
-  const key_counts: { [key: string]: number } = {}
+export function keyCounts(objs: any[], depth = 0) {
+  const key_counts: {[key: string]: number} = {};
 
   for (const obj of objs) {
-    let Q = ObjectItems(obj) as [string, any][]
-    let d = depth
+    let Q = ObjectItems(obj) as [string, any][];
+    const d = depth;
     while (Q.length > 0) {
-      let [k, v] = Q.pop() as [string, any]
+      const [k, v] = Q.pop() as [string, any];
 
-      if (key_counts[k] === undefined)
-        key_counts[k] = 0
-      key_counts[k] += 1
-
+      if (key_counts[k] === undefined) key_counts[k] = 0;
+      key_counts[k] += 1;
 
       if (k.replace(/[^\.]/g, '').length < d) {
         try {
@@ -27,23 +25,21 @@ export function keyCounts(objs: any[], depth: number = 0) {
             Q = Q.concat(
               ObjectItems(v).map(
                 ([kk, vv]) =>
-                  [
-                    [k, kk].filter(
-                      (kkk) => kkk !== ''
-                    ).join('.'),
-                    vv
-                  ] as [string, any]
-              ) as [string, any][]
-            )
+                  [[k, kk].filter(kkk => kkk !== '').join('.'), vv] as [
+                    string,
+                    any,
+                  ],
+              ) as [string, any][],
+            );
           }
         } catch (e) {
-          debug(e)
+          debug(e);
         }
       }
     }
   }
 
-  return key_counts
+  return key_counts;
 }
 
 /**
@@ -52,28 +48,24 @@ export function keyCounts(objs: any[], depth: number = 0) {
  * @param objs The objects to enumerate the values of
  * @param depth The depth to go into object keys
  */
-export function valueCounts(objs: any[], depth: number = 0) {
-  const key_value_counts: { [key: string]: { [value: string]: number } } = {}
+export function valueCounts(objs: any[], depth = 0) {
+  const key_value_counts: {[key: string]: {[value: string]: number}} = {};
 
   for (const obj of objs) {
-    let Q = ObjectItems(obj)
-    let d = depth
+    let Q = ObjectItems(obj);
+    const d = depth;
     while (Q.length > 0) {
-      let [k, v] = Q.pop() as [string, any]
+      const [k, v] = Q.pop() as [string, any];
 
-      let v_str
-      if (Array.isArray(v))
-        v_str = '[array]'
-      else if (typeof v === 'object')
-        v_str = '[object]'
-      else
-        v_str = v + ''
+      let v_str;
+      if (Array.isArray(v)) v_str = '[array]';
+      else if (typeof v === 'object') v_str = '[object]';
+      else v_str = v + '';
 
-      if (key_value_counts[k] === undefined)
-        key_value_counts[k] = {}
+      if (key_value_counts[k] === undefined) key_value_counts[k] = {};
       if (key_value_counts[k][v_str] === undefined)
-        key_value_counts[k][v_str] = 0
-      key_value_counts[k][v_str] += 1
+        key_value_counts[k][v_str] = 0;
+      key_value_counts[k][v_str] += 1;
 
       if (k.replace(/[^\.]/g, '').length < d) {
         try {
@@ -81,19 +73,17 @@ export function valueCounts(objs: any[], depth: number = 0) {
             Q = Q.concat(
               ObjectItems(v).map(
                 ([kk, vv]) =>
-                  [
-                    [k, kk].filter(
-                      (kkk) => kkk !== ''
-                    ).join('.'),
-                    vv
-                  ] as [string, any]
-              ) as [string, any][]
-            )
+                  [[k, kk].filter(kkk => kkk !== '').join('.'), vv] as [
+                    string,
+                    any,
+                  ],
+              ) as [string, any][],
+            );
           }
-        } catch (e) { }
+        } catch (e) {}
       }
     }
   }
 
-  return key_value_counts
+  return key_value_counts;
 }
