@@ -1,3 +1,3 @@
 export function sum(nums: number[]) {
-  return nums.reduce((total, num) => total + num)
+  return nums.reduce((total, num) => total + num);
 }
