@@ -1,5 +1,5 @@
-import * as assert from 'assert'
-import { makeTemplate } from '../../../src/util/dynamic-template'
+import * as assert from 'assert';
+import {makeTemplate} from '../../../src/util/dynamic-template';
 
 describe('util', () => {
   describe('dynamic-template', () => {
@@ -7,21 +7,18 @@ describe('util', () => {
       it('works with no args', () => {
         assert.equal(
           makeTemplate('hello test', {me: 'test', you: 'no'}),
-          'hello test'
-        )
-      })
+          'hello test',
+        );
+      });
       it('works', () => {
-        assert.equal(
-          makeTemplate('hello ${me}', {me: 'test'}),
-          'hello test'
-        )
-      })
+        assert.equal(makeTemplate('hello ${me}', {me: 'test'}), 'hello test');
+      });
       it('works with too many args', () => {
         assert.equal(
           makeTemplate('hello ${me}', {me: 'test', you: 'no'}),
-          'hello test'
-        )
-      })
-    })
-  })
-})
+          'hello test',
+        );
+      });
+    });
+  });
+});
