@@ -1,5 +1,5 @@
-import * as assert from 'assert'
-import { flatten_keys } from '../../../src/util/flatten-keys'
+import * as assert from 'assert';
+import {flatten_keys} from '../../../src/util/flatten-keys';
 
 describe('util', () => {
   describe('flatten_keys', () => {
@@ -10,22 +10,22 @@ describe('util', () => {
           meta: {
             test: {
               me: ['plenty', 'always'],
-              and: 2
+              and: 2,
             },
             best: {
               nest: {
-                guest: '3'
-              }
-            }
-          }
+                guest: '3',
+              },
+            },
+          },
         }),
         {
           id: '1',
           'meta.test.me': ['plenty', 'always'],
           'meta.test.and': 2,
-          'meta.best.nest.guest': '3'
-        }
-      )
-    })
-  })
-})
+          'meta.best.nest.guest': '3',
+        },
+      );
+    });
+  });
+});
